Add files via upload
Data analysis has proven to be very effective in sales. This is because we can extract key performance indicators (KPIs) to make smarter decisions. This saves time, resources, and money. Sales data analysis can cover many aspects (sales, production, customer experience, employee efficiency); therefore, it presents many challenges. The goal of this project is to analyze data from a sales records database for scale model cars and extract information for decision-making.

Good analysis starts with questions. Below are the questions that this projects provides the answers for.

Question 1: Which products should we order more of or less of?
Question 2: How should we tailor marketing and communication strategies to customer behaviors?
Question 3: How much can we spend on acquiring new customers?